edited 5-sqrt_recursion.c

+++ 0x08-recursion/5-sqrt_recursion.c
@@ -14,7 +14,7 @@ int find_sqrt(int n, int start, int end)
 		return (-1);
 
 	int mid = start + (end - start) / 2;
-	long long square = (long long)mid * mid;
+	int square = mid * mid;
 
 	if (square == n)
 		return (mid);
